fix: handle unset appmap.output.directory
Make sure Properties.OutputDirectory is set correctly when the config
contains appmap_dir and system property appmap.output.directory is
unset. Prior to these changes, OutputDirectory was getting set to a
default value, rather than what the config contained.
