+++ agent/test/spark/app/appmap.yml
@@ -1,4 +1,4 @@
 name: sparkexample
-appmap_dir: app/build/tmp/appmap
+appmap_dir: build/tmp/appmap
 packages:
 - path: com.appland.appmap.sparkexample
